feat(R-0001): sub-8 — thin binaries + tower-sessions cookies + utoipa OpenAPI + scoped CORS
Promote each `bin/*/src/main.rs` runtime out of the binary crate and into a
new per-binary library crate (`crates/tanren-{api,cli,mcp,tui}-app/`) per the
thin-binary-crate profile. The binaries shrink to ≤ 50-line wiring shells
that initialize tracing via `tanren_observability::init(...)` and hand off
to the app crate; `just check-thin-binary` and `xtask check-tracing-init`
are wired into `just check` and the lefthook pre-commit suite.

`tanren-api-app` lands the full sub-8 boundary surface:

- **Configurable CORS.** `Config::cors_allow_origins: Vec<HeaderValue>`
  reads `TANREN_API_CORS_ORIGINS` (comma-separated); default
  `http://localhost:3000`. `tower_http::cors::Any` is denied
  workspace-wide.
- **Cookie sessions.** `tower-sessions = "0.14"` +
  `tower-sessions-sqlx-store = "0.15"` (sqlite + postgres). Successful
  sign-up/sign-in/accept-invitation responses set `Set-Cookie:
  tanren_session=<id>; HttpOnly; Secure; SameSite=Strict; Path=/;
  Max-Age=2592000` and return `SessionEnvelope::Cookie`. The new
  `POST /sessions/revoke` route flushes the session via
  `Session::flush` and returns 204. The cookie store and the legacy
  `account_sessions` table coexist — bearer-flow callers
  (CLI/MCP/TUI) keep their `SessionView` body shape unchanged.
- **utoipa OpenAPI.** Replaces the hand-rolled
  `serde_json::json!({...})` document. Each handler carries
  `#[utoipa::path(...)]`; the top-level `ApiDoc` derives `OpenApi`,
  and `OpenApiRouter` from `utoipa-axum` serves `/openapi.json`.
  Contract types in `tanren-contract` and newtype ids in
  `tanren-identity-policy` dual-derive `JsonSchema` + `ToSchema` so
  rmcp's tool registry and utoipa's OpenAPI document share the same
  source of truth.

`tanren_observability::init` now takes an explicit `default_filter`
argument; `RUST_LOG` continues to override when set. All five binaries
(`tanren-api`, `tanren-cli`, `tanren-mcp`, `tanren-tui`, `tanrend`)
call `tanren_observability::init(tanren_observability::default_filter())`
first thing in `main`.

`just check-deps` is updated to also accept the axum dependency on
`tanren-mcp-app` (the runtime crate) — the bin shell no longer depends
on axum directly.
