+++ bin/tanren-api/Cargo.toml
@@ -13,14 +13,6 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-axum = { workspace = true }
-secrecy = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tanren-app-services = { path = "../../crates/tanren-app-services" }
-tanren-contract = { path = "../../crates/tanren-contract" }
-tanren-identity-policy = { path = "../../crates/tanren-identity-policy" }
+tanren-api-app = { path = "../../crates/tanren-api-app" }
 tanren-observability = { path = "../../crates/tanren-observability" }
 tokio = { workspace = true }
-tower-http = { workspace = true }
-tracing = { workspace = true }
