+++ bin/tanren-cli/src/main.rs
@@ -1,299 +1,18 @@
 //! Tanren scriptable command-line client.
 //!
-//! R-0001 (S-06) adds the cross-interface `account` subcommand
-//! (`create`, `sign-in`) backed by the same `tanren-app-services`
-//! handlers the api / mcp / tui / web surfaces use. Sessions are
-//! persisted at `$TANREN_SESSION_FILE` (default
-//! `$XDG_STATE_HOME/tanren/session`) so the next invocation can
-//! verify the session round-trips.
+//! Thin entry point per
+//! `profiles/rust-cargo/architecture/thin-binary-crate.md`. All runtime
+//! logic lives in `tanren-cli-app`; this `main` parses the CLI,
+//! initializes tracing, and hands off.
 
-use std::env;
-use std::fs;
-use std::io::Write;
-use std::path::PathBuf;
 use std::process::ExitCode;
 
 use anyhow::{Context, Result};
-use clap::{Parser, Subcommand};
-use secrecy::SecretString;
-use tanren_app_services::{AppServiceError, Handlers, Store};
-use tanren_contract::{AcceptInvitationRequest, SignInRequest, SignUpRequest};
-use tanren_identity_policy::{Email, InvitationToken};
+use tanren_cli_app::{Config, run};
 
-const SESSION_FILE_ENV: &str = "TANREN_SESSION_FILE";
-
-#[derive(Debug, Parser)]
-#[command(
-    name = "tanren-cli",
-    version,
-    about = "Tanren scriptable command-line client"
-)]
-struct Cli {
-    #[command(subcommand)]
-    command: Option<Command>,
-}
-
-#[derive(Debug, Subcommand)]
-enum Command {
-    /// Print a liveness report. Mirrors the api `/health` endpoint.
-    Health,
-    /// Database migration commands.
-    Migrate {
-        #[command(subcommand)]
-        action: MigrateAction,
-    },
-    /// Account flow: self-signup, sign-in, accept-invitation.
-    Account {
-        #[command(subcommand)]
-        action: AccountAction,
-    },
-}
-
-#[derive(Debug, Subcommand)]
-enum MigrateAction {
-    /// Apply all pending migrations.
-    Up {
-        /// Database URL (defaults to `$DATABASE_URL`).
-        #[arg(long, env = "DATABASE_URL")]
-        database_url: String,
-    },
-}
-
-#[derive(Debug, Subcommand)]
-enum AccountAction {
-    /// Create a personal account (or, with `--invitation`, accept an
-    /// invitation and join the inviting org).
-    Create {
-        /// Database URL.
-        #[arg(long, env = "DATABASE_URL")]
-        database_url: String,
-        /// Email to register.
-        #[arg(long)]
-        identifier: String,
-        /// Password.
-        #[arg(long)]
-        password: String,
-        /// Display name.
-        #[arg(long, default_value_t = String::from("Tanren user"))]
-        display_name: String,
-        /// Optional invitation token. When supplied, the new account
-        /// joins the inviting org instead of being a personal account.
-        #[arg(long)]
-        invitation: Option<String>,
-    },
-    /// Sign in to an existing account and persist the session.
-    SignIn {
-        /// Database URL.
-        #[arg(long, env = "DATABASE_URL")]
-        database_url: String,
-        /// Email to sign in with.
-        #[arg(long)]
-        identifier: String,
-        /// Password.
-        #[arg(long)]
-        password: String,
-    },
-}
-
-fn main() -> ExitCode {
-    let cli = Cli::parse();
-    let result = match cli.command {
-        None | Some(Command::Health) => print_health(),
-        Some(Command::Migrate {
-            action: MigrateAction::Up { database_url },
-        }) => run_migrate_up(&database_url),
-        Some(Command::Account { action }) => dispatch_account(action),
-    };
-    match result {
-        Ok(()) => ExitCode::SUCCESS,
-        Err(err) => {
-            let stderr = std::io::stderr();
-            let mut handle = stderr.lock();
-            let _ = writeln!(handle, "{err}");
-            ExitCode::from(1)
-        }
-    }
-}
-
-fn print_health() -> Result<()> {
-    let report = Handlers::new().health(env!("CARGO_PKG_VERSION"));
-    let stdout = std::io::stdout();
-    let mut handle = stdout.lock();
-    writeln!(
-        handle,
-        "status={status} version={version} contract_version={contract}",
-        status = report.status,
-        version = report.version,
-        contract = report.contract_version.value(),
-    )
-    .context("write health report to stdout")?;
-    Ok(())
-}
-
-fn run_migrate_up(database_url: &str) -> Result<()> {
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("build tokio runtime")?;
-    runtime.block_on(async {
-        Handlers::new()
-            .migrate(database_url)
-            .await
-            .context("apply pending migrations")
-    })?;
-    let stdout = std::io::stdout();
-    let mut handle = stdout.lock();
-    writeln!(handle, "migrations: applied").context("write migrate report to stdout")?;
-    Ok(())
-}
-
-fn dispatch_account(action: AccountAction) -> Result<()> {
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("build tokio runtime")?;
-    runtime.block_on(run_account(action))
-}
-
-async fn run_account(action: AccountAction) -> Result<()> {
-    let handlers = Handlers::new();
-    match action {
-        AccountAction::Create {
-            database_url,
-            identifier,
-            password,
-            display_name,
-            invitation,
-        } => {
-            let store = Store::connect(&database_url)
-                .await
-                .context("connect to store")?;
-            let email = Email::parse(&identifier).context("parse --identifier as email")?;
-            let password = SecretString::from(password);
-            match invitation {
-                None => {
-                    let response = handlers
-                        .sign_up(
-                            &store,
-                            SignUpRequest {
-                                email,
-                                password,
-                                display_name,
-                            },
-                        )
-                        .await
-                        .map_err(account_error)?;
-                    persist_session(response.session.token.expose_secret())?;
-                    let stdout = std::io::stdout();
-                    let mut handle = stdout.lock();
-                    writeln!(
-                        handle,
-                        "account_id={id} session={token}",
-                        id = response.account.id,
-                        token = response.session.token.expose_secret(),
-                    )
-                    .context("write sign-up result")?;
-                }
-                Some(token) => {
-                    let invitation_token = InvitationToken::parse(&token)
-                        .context("parse --invitation as invitation token")?;
-                    let response = handlers
-                        .accept_invitation(
-                            &store,
-                            AcceptInvitationRequest {
-                                invitation_token,
-                                email,
-                                password,
-                                display_name,
-                            },
-                        )
-                        .await
-                        .map_err(account_error)?;
-                    persist_session(response.session.token.expose_secret())?;
-                    let stdout = std::io::stdout();
-                    let mut handle = stdout.lock();
-                    writeln!(
-                        handle,
-                        "account_id={id} session={token} joined_org={org}",
-                        id = response.account.id,
-                        token = response.session.token.expose_secret(),
-                        org = response.joined_org,
-                    )
-                    .context("write invitation-acceptance result")?;
-                }
-            }
-        }
-        AccountAction::SignIn {
-            database_url,
-            identifier,
-            password,
-        } => {
-            let store = Store::connect(&database_url)
-                .await
-                .context("connect to store")?;
-            let email = Email::parse(&identifier).context("parse --identifier as email")?;
-            let password = SecretString::from(password);
-            let response = handlers
-                .sign_in(&store, SignInRequest { email, password })
-                .await
-                .map_err(account_error)?;
-            persist_session(response.session.token.expose_secret())?;
-            let stdout = std::io::stdout();
-            let mut handle = stdout.lock();
-            writeln!(
-                handle,
-                "account_id={id} session={token}",
-                id = response.account.id,
-                token = response.session.token.expose_secret(),
-            )
-            .context("write sign-in result")?;
-        }
-    }
-    Ok(())
-}
-
-fn account_error(err: AppServiceError) -> anyhow::Error {
-    match err {
-        AppServiceError::Account(reason) => {
-            anyhow::anyhow!("error: {} — {}", reason.code(), reason.summary())
-        }
-        AppServiceError::InvalidInput(message) => {
-            anyhow::anyhow!("error: validation_failed — {message}")
-        }
-        AppServiceError::Store(err) => {
-            anyhow::anyhow!("error: internal_error — {err}")
-        }
-        _ => anyhow::anyhow!("error: internal_error — unknown app-service failure"),
-    }
-}
-
-fn session_path() -> PathBuf {
-    if let Ok(explicit) = env::var(SESSION_FILE_ENV) {
-        if !explicit.is_empty() {
-            return PathBuf::from(explicit);
-        }
-    }
-    let base = env::var("XDG_STATE_HOME")
-        .ok()
-        .filter(|s| !s.is_empty())
-        .map_or_else(
-            || {
-                env::var("HOME").ok().map_or_else(
-                    || PathBuf::from("."),
-                    |home| PathBuf::from(home).join(".local/state"),
-                )
-            },
-            PathBuf::from,
-        );
-    base.join("tanren").join("session")
-}
-
-fn persist_session(token: &str) -> Result<()> {
-    let path = session_path();
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("create session dir {}", parent.display()))?;
-    }
-    fs::write(&path, token).with_context(|| format!("write session to {}", path.display()))?;
-    Ok(())
+fn main() -> Result<ExitCode> {
+    tanren_observability::init(tanren_observability::default_filter())
+        .context("install tracing subscriber")?;
+    let config = Config::parse_from_env();
+    Ok(run(config))
 }
