+++ bin/tanren-mcp/Cargo.toml
@@ -13,16 +13,6 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-axum = { workspace = true }
-rmcp = { workspace = true }
-secrecy = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tanren-app-services = { path = "../../crates/tanren-app-services" }
-tanren-contract = { path = "../../crates/tanren-contract" }
+tanren-mcp-app = { path = "../../crates/tanren-mcp-app" }
 tanren-observability = { path = "../../crates/tanren-observability" }
 tokio = { workspace = true }
-tokio-util = { workspace = true }
-tower = { workspace = true }
-tower-http = { workspace = true }
-tracing = { workspace = true }
