+++ bin/tanren-mcp/src/main.rs
@@ -1,434 +1,17 @@
 //! Tanren MCP (Model Context Protocol) server.
 //!
-//! F-0002 corrects F-0001's stdio scaffolding: the architecture mandates
-//! MCP Streamable HTTP with scoped credentials (see
-//! `docs/architecture/subsystems/interfaces.md#mcp` and
-//! `docs/architecture/technology.md`). This binary wraps an `rmcp`
-//! `StreamableHttpService` in `axum`, gates `/mcp` behind an API-key
-//! middleware that emits the shared
-//! `auth_required` / `permission_denied` taxonomy from the interfaces
-//! contract, and exposes the same `/health` shape as `tanren-api`.
-//!
-//! R-0001 (S-07) registers the first three account-flow tools
-//! (`account.create`, `account.sign_in`, `account.accept_invitation`)
-//! via `#[rmcp::tool]` routes that delegate to `tanren-app-services`.
+//! Thin entry point per
+//! `profiles/rust-cargo/architecture/thin-binary-crate.md`. All runtime
+//! logic — rmcp tool registry, API-key middleware, host-header
+//! allowlist, streamable-HTTP server — lives in `tanren-mcp-app`; this
+//! `main` initializes tracing and hands off.
 
 use anyhow::{Context, Result};
-use axum::Json;
-use axum::Router;
-use axum::extract::Request;
-use axum::http::{HeaderMap, StatusCode, header};
-use axum::middleware::{self, Next};
-use axum::response::{IntoResponse, Response};
-use axum::routing::get;
-use rmcp::ErrorData as McpError;
-use rmcp::ServerHandler;
-use rmcp::handler::server::router::tool::ToolRouter;
-use rmcp::handler::server::wrapper::Parameters;
-use rmcp::model::{CallToolResult, Content, ServerCapabilities, ServerInfo};
-use rmcp::transport::streamable_http_server::{
-    StreamableHttpServerConfig, StreamableHttpService, session::local::LocalSessionManager,
-};
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-use std::env;
-use std::sync::Arc;
-use tanren_app_services::{AppServiceError, Handlers, Store};
-use tanren_contract::{AcceptInvitationRequest, SignInRequest, SignUpRequest};
-use tokio::net::TcpListener;
-use tokio_util::sync::CancellationToken;
-use tower::ServiceBuilder;
-use tower_http::cors::{Any, CorsLayer};
-
-const DEFAULT_BIND_ADDRESS: &str = "0.0.0.0:8081";
-const BIND_ADDRESS_ENV: &str = "TANREN_MCP_BIND";
-const API_KEY_ENV: &str = "TANREN_MCP_API_KEY";
-const DATABASE_URL_ENV: &str = "DATABASE_URL";
-/// Comma-separated extra hostnames / `host:port` authorities to add to
-/// rmcp's `allowed_hosts` Host-header allowlist.
-const ALLOWED_HOSTS_ENV: &str = "TANREN_MCP_ALLOWED_HOSTS";
-
-/// MCP tool surface. Holds the shared `Handlers` facade and a `Store`
-/// handle; behaviour tools delegate through the facade so the api / mcp /
-/// cli / tui surfaces all resolve to the same logic per the
-/// equivalent-operations rule in
-/// `docs/architecture/subsystems/interfaces.md`.
-#[derive(Clone)]
-struct TanrenMcp {
-    handlers: Handlers,
-    store: Arc<Store>,
-    /// Cached tool router built from the `#[rmcp::tool]` methods on this
-    /// type. Read by the macro-generated `ServerHandler` impl below.
-    tool_router: ToolRouter<Self>,
-}
-
-impl std::fmt::Debug for TanrenMcp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TanrenMcp").finish_non_exhaustive()
-    }
-}
-
-#[rmcp::tool_router]
-impl TanrenMcp {
-    fn new(handlers: Handlers, store: Arc<Store>) -> Self {
-        Self {
-            handlers,
-            store,
-            tool_router: Self::tool_router(),
-        }
-    }
-
-    /// Self-signup tool. Mirrors the api `POST /accounts` shape via
-    /// `tanren_contract::SignUpRequest` / `SignUpResponse`.
-    #[rmcp::tool(
-        name = "account.create",
-        description = "Create a new Tanren account via self-signup. Returns the new account view and an opaque session token. Failures use the shared {code, summary} taxonomy: duplicate_identifier, invalid_credential."
-    )]
-    async fn account_create(
-        &self,
-        Parameters(request): Parameters<SignUpRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        match self.handlers.sign_up(self.store.as_ref(), request).await {
-            Ok(response) => Ok(success(&response)),
-            Err(err) => Ok(map_failure(err)),
-        }
-    }
-
-    /// Sign-in tool. Mirrors the api `POST /sessions` shape via
-    /// `tanren_contract::SignInRequest` / `SignInResponse`.
-    #[rmcp::tool(
-        name = "account.sign_in",
-        description = "Sign in to an existing Tanren account. Returns the account view and an opaque session token. Failure code: invalid_credential."
-    )]
-    async fn account_sign_in(
-        &self,
-        Parameters(request): Parameters<SignInRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        match self.handlers.sign_in(self.store.as_ref(), request).await {
-            Ok(response) => Ok(success(&response)),
-            Err(err) => Ok(map_failure(err)),
-        }
-    }
-
-    /// Invitation-acceptance tool. Mirrors the api
-    /// `POST /invitations/{token}/accept` shape via
-    /// `tanren_contract::AcceptInvitationRequest` /
-    /// `AcceptInvitationResponse`.
-    #[rmcp::tool(
-        name = "account.accept_invitation",
-        description = "Accept an organization invitation and create a Tanren account in the inviting org. Failure codes: invitation_not_found, invitation_already_consumed, invitation_expired, invalid_credential."
-    )]
-    async fn account_accept_invitation(
-        &self,
-        Parameters(request): Parameters<AcceptInvitationRequest>,
-    ) -> Result<CallToolResult, McpError> {
-        match self
-            .handlers
-            .accept_invitation(self.store.as_ref(), request)
-            .await
-        {
-            Ok(response) => Ok(success(&response)),
-            Err(err) => Ok(map_failure(err)),
-        }
-    }
-
-    /// Borrow the cached `ToolRouter`. Exists so the dead-code lint can
-    /// see the field as read even on rmcp macro versions whose
-    /// `#[tool_handler]` expansion path does not access the field
-    /// directly under the lint's heuristic. Production callers reach
-    /// the router via the `ServerHandler` trait's `call_tool` /
-    /// `list_tools` methods generated by `#[tool_handler]`, not this
-    /// helper.
-    fn router(&self) -> &ToolRouter<Self> {
-        &self.tool_router
-    }
-}
-
-#[rmcp::tool_handler]
-impl ServerHandler for TanrenMcp {
-    fn get_info(&self) -> ServerInfo {
-        // Touch the cached router so the dead-code lint never flags
-        // `tool_router` even on rmcp macro versions whose tool_handler
-        // expansion path uses the static `Self::tool_router()` builder
-        // rather than the cached field.
-        let _ = self.router();
-        let mut info = ServerInfo::default();
-        info.instructions = Some(
-            "Tanren control plane MCP server. Account-flow tools route through the same handlers the HTTP API uses; failure responses share the {code, summary} error taxonomy."
-                .to_owned(),
-        );
-        info.capabilities = ServerCapabilities::builder().enable_tools().build();
-        info
-    }
-}
-
-/// Encode a successful handler response as a JSON-text `CallToolResult`.
-fn success<T: Serialize>(value: &T) -> CallToolResult {
-    let text = serde_json::to_string(value).unwrap_or_else(|_| "{}".to_owned());
-    CallToolResult::success(vec![Content::text(text)])
-}
-
-/// Encode an [`AppServiceError`] as the shared `{code, summary}` error
-/// body and surface it as an MCP tool failure result.
-fn map_failure(err: AppServiceError) -> CallToolResult {
-    let (code, summary) = match err {
-        AppServiceError::Account(reason) => (reason.code().to_owned(), reason.summary().to_owned()),
-        AppServiceError::InvalidInput(message) => ("validation_failed".to_owned(), message),
-        AppServiceError::Store(err) => (
-            "internal_error".to_owned(),
-            format!("Tanren encountered an internal error: {err}"),
-        ),
-        _ => (
-            "internal_error".to_owned(),
-            "Unknown app-service failure".to_owned(),
-        ),
-    };
-    let body = json!({
-        "code": code,
-        "summary": summary,
-    });
-    let text = serde_json::to_string(&body).unwrap_or_else(|_| "{}".to_owned());
-    CallToolResult::error(vec![Content::text(text)])
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct HealthResponse {
-    status: String,
-    version: String,
-    contract_version: u32,
-}
-
-async fn health() -> Json<HealthResponse> {
-    let report = Handlers::new().health(env!("CARGO_PKG_VERSION"));
-    Json(HealthResponse {
-        status: report.status.to_owned(),
-        version: report.version.to_owned(),
-        contract_version: report.contract_version.value(),
-    })
-}
-
-/// Shared error response shape per
-/// `docs/architecture/subsystems/interfaces.md` "Error Taxonomy".
-fn error_body(code: &str, summary: &str) -> serde_json::Value {
-    json!({
-        "code": code,
-        "summary": summary,
-    })
-}
-
-#[derive(Debug, Clone)]
-struct AuthConfig {
-    /// Bootstrap API key. F-0002 sources this from `TANREN_MCP_API_KEY`;
-    /// R-0008 will route through the real credential store. Wrapped in
-    /// `SecretString` so accidental `Debug` / `Serialize` calls do not
-    /// leak the credential.
-    bootstrap_key: Option<secrecy::SecretString>,
-}
-
-impl AuthConfig {
-    fn from_env() -> Self {
-        let bootstrap_key = env::var(API_KEY_ENV)
-            .ok()
-            .filter(|s| !s.is_empty())
-            .map(secrecy::SecretString::from);
-        Self { bootstrap_key }
-    }
-
-    fn extract_credential(headers: &HeaderMap) -> Option<&str> {
-        if let Some(value) = headers
-            .get(header::AUTHORIZATION)
-            .and_then(|v| v.to_str().ok())
-            && let Some(token) = value
-                .strip_prefix("Bearer ")
-                .or_else(|| value.strip_prefix("bearer "))
-        {
-            return Some(token.trim());
-        }
-        if let Some(value) = headers.get("x-api-key").and_then(|v| v.to_str().ok()) {
-            return Some(value.trim());
-        }
-        None
-    }
-}
-
-async fn require_api_key(
-    axum::extract::State(config): axum::extract::State<Arc<AuthConfig>>,
-    request: Request,
-    next: Next,
-) -> Response {
-    // Operator-config check first: an unconfigured server is in an
-    // outage state, not an auth-failure state.
-    let Some(expected) = config
-        .bootstrap_key
-        .as_ref()
-        .map(secrecy::ExposeSecret::expose_secret)
-    else {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(error_body(
-                "unavailable",
-                "MCP credential store is not configured. Set TANREN_MCP_API_KEY (bootstrap key) until R-0008 lands the real store.",
-            )),
-        )
-            .into_response();
-    };
-
-    let Some(presented) = AuthConfig::extract_credential(request.headers()) else {
-        return (
-            StatusCode::UNAUTHORIZED,
-            Json(error_body(
-                "auth_required",
-                "Missing Authorization: Bearer <api-key> or X-API-Key header.",
-            )),
-        )
-            .into_response();
-    };
-
-    if !constant_time_eq(presented.as_bytes(), expected.as_bytes()) {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(error_body(
-                "permission_denied",
-                "Presented credential is not authorized for this MCP service.",
-            )),
-        )
-            .into_response();
-    }
-
-    next.run(request).await
-}
-
-fn constant_time_eq(a: &[u8], b: &[u8]) -> bool {
-    if a.len() != b.len() {
-        return false;
-    }
-    let mut diff: u8 = 0;
-    for (x, y) in a.iter().zip(b.iter()) {
-        diff |= x ^ y;
-    }
-    diff == 0
-}
-
-fn build_router(
-    auth_config: Arc<AuthConfig>,
-    handlers: Handlers,
-    store: Arc<Store>,
-    cancellation: CancellationToken,
-) -> Router {
-    let config = streamable_http_config(cancellation);
-    let mcp_service: StreamableHttpService<TanrenMcp, LocalSessionManager> =
-        StreamableHttpService::new(
-            move || Ok(TanrenMcp::new(handlers.clone(), store.clone())),
-            Arc::new(LocalSessionManager::default()),
-            config,
-        );
-
-    let mcp_with_auth = ServiceBuilder::new()
-        .layer(middleware::from_fn_with_state(auth_config, require_api_key))
-        .service(mcp_service);
-
-    let cors = CorsLayer::new()
-        .allow_origin(Any)
-        .allow_methods(Any)
-        .allow_headers(Any);
-
-    Router::new()
-        .route("/health", get(health))
-        .nest_service("/mcp", mcp_with_auth)
-        .layer(cors)
-}
-
-/// Build rmcp's `StreamableHttpServerConfig` honouring the
-/// `TANREN_MCP_ALLOWED_HOSTS` env var.
-fn streamable_http_config(cancellation: CancellationToken) -> StreamableHttpServerConfig {
-    let base = StreamableHttpServerConfig::default().with_cancellation_token(cancellation);
-    let raw = env::var(ALLOWED_HOSTS_ENV).ok().filter(|s| !s.is_empty());
-    let Some(value) = raw else {
-        return base;
-    };
-    if value.trim() == "*" {
-        tracing::warn!(
-            target: "tanren_mcp",
-            env_var = ALLOWED_HOSTS_ENV,
-            "Host-header validation disabled by `*`; relying on API-key auth as the sole gate."
-        );
-        return base.disable_allowed_hosts();
-    }
-    let mut hosts: Vec<String> = vec!["localhost".into(), "127.0.0.1".into(), "::1".into()];
-    for host in value.split(',') {
-        let trimmed = host.trim();
-        if !trimmed.is_empty() {
-            hosts.push(trimmed.to_owned());
-        }
-    }
-    tracing::info!(
-        target: "tanren_mcp",
-        allowed_hosts = ?hosts,
-        "Host-header validation extended via {ALLOWED_HOSTS_ENV}"
-    );
-    base.with_allowed_hosts(hosts)
-}
+use tanren_mcp_app::{Config, serve};
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    tanren_observability::init().context("install tracing subscriber")?;
-
-    let bind = env::var(BIND_ADDRESS_ENV).unwrap_or_else(|_| DEFAULT_BIND_ADDRESS.to_owned());
-    let auth_config = Arc::new(AuthConfig::from_env());
-    if auth_config.bootstrap_key.is_none() {
-        tracing::warn!(
-            target: "tanren_mcp",
-            env_var = API_KEY_ENV,
-            "TANREN_MCP_API_KEY is not set — every /mcp request will be rejected with `unavailable` until a bootstrap key is provided."
-        );
-    }
-
-    let database_url = env::var(DATABASE_URL_ENV).with_context(|| {
-        format!("{DATABASE_URL_ENV} must be set so tanren-mcp can connect to the event store")
-    })?;
-    let store = Arc::new(
-        Store::connect(&database_url)
-            .await
-            .with_context(|| format!("connect to store at {DATABASE_URL_ENV}"))?,
-    );
-    let handlers = Handlers::new();
-
-    let cancellation = CancellationToken::new();
-    let router = build_router(auth_config, handlers, store, cancellation.clone());
-
-    let listener = TcpListener::bind(&bind)
-        .await
-        .with_context(|| format!("bind {bind}"))?;
-    tracing::info!(target: "tanren_mcp", address = %bind, "tanren-mcp listening on streamable HTTP");
-
-    let cancel = cancellation.clone();
-    axum::serve(listener, router)
-        .with_graceful_shutdown(async move {
-            shutdown_signal().await;
-            cancel.cancel();
-        })
-        .await
-        .context("axum serve")?;
-    Ok(())
-}
-
-#[cfg(unix)]
-async fn shutdown_signal() {
-    use tokio::signal::unix::{SignalKind, signal};
-    let sigterm = signal(SignalKind::terminate()).ok();
-    if let Some(mut sigterm) = sigterm {
-        tokio::select! {
-            _ = tokio::signal::ctrl_c() => {}
-            _ = sigterm.recv() => {}
-        }
-    } else {
-        let _ = tokio::signal::ctrl_c().await;
-    }
-    tracing::info!(target: "tanren_mcp", "shutdown signal received");
-}
-
-#[cfg(not(unix))]
-async fn shutdown_signal() {
-    let _ = tokio::signal::ctrl_c().await;
-    tracing::info!(target: "tanren_mcp", "shutdown signal received");
+    tanren_observability::init(tanren_observability::default_filter())
+        .context("install tracing subscriber")?;
+    serve(Config::from_env()).await
 }
