+++ crates/tanren-cli-app/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "tanren-cli-app"
+version = "0.0.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish = false
+description = "Library crate that hosts the tanren-cli runtime — argument parsing, account-flow dispatch, session persistence. The bin/tanren-cli/src/main.rs entry point is a thin wiring shell that initializes tracing and hands off to this crate."
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+clap = { workspace = true }
+secrecy = { workspace = true }
+tanren-app-services = { path = "../tanren-app-services" }
+tanren-contract = { path = "../tanren-contract" }
+tanren-identity-policy = { path = "../tanren-identity-policy" }
+tokio = { workspace = true }
