+++ crates/tanren-mcp-app/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "tanren-mcp-app"
+version = "0.0.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish = false
+description = "Library crate that hosts the tanren-mcp runtime — rmcp tool surface, API-key middleware, host-header allowlist, and the streamable-HTTP server. The bin/tanren-mcp/src/main.rs entry point is a thin wiring shell."
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+axum = { workspace = true }
+rmcp = { workspace = true }
+secrecy = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tanren-app-services = { path = "../tanren-app-services" }
+tanren-contract = { path = "../tanren-contract" }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+tower = { workspace = true }
+tower-http = { workspace = true }
+tracing = { workspace = true }
