+++ crates/tanren-tui-app/src/draw.rs
